+++ library/common/bridge/utility.h
@@ -5,9 +5,9 @@
 #include "library/common/types/c_types.h"
 
 namespace Envoy {
-namespace Types {
+namespace Bridge {
 
 envoy_map makeEnvoyMap(std::vector<std::pair<std::string, std::string>> pairs);
 
-} // namespace Types
+} // namespace Bridge
 } // namespace Envoy
