+++ library/common/extensions/filters/http/test_event_tracker/BUILD
@@ -26,8 +26,8 @@ envoy_cc_extension(
         "filter_cc_proto",
         "//library/common/api:c_types",
         "//library/common/api:external_api_lib",
+        "//library/common/bridge:utility_lib",
         "//library/common/data:utility_lib",
-        "//library/common/types:utility_lib",
         "@envoy//source/common/common:assert_lib",
         "@envoy//source/extensions/filters/http/common:pass_through_filter_lib",
     ],
